fix: cleanup dead code and improve robustness
- Removed unused `context/webSocketReducer.ts` and migrated logic to `context/WebSocketContext.tsx`.
- Updated `WebSocketContext.tsx` to handle `INITIAL_STATE` and reducer logic locally.
- Added explanatory comment to `hooks/useSpotifyWebPlayback.ts` regarding `useSpotifyAuth` reliance.
- Fixed imports in `types/websocket.ts` and `WebSocketContext.tsx` to ensure type safety.
- Updated `tests/unit/context/webSocketReducer.test.ts` to import from `WebSocketContext`.
